Move device enumeration into the rmaps base
The device enumeration lived as file statics in round_robin because that
was the only mapper that needed it.  It is about to have a second caller:
"ppr:N:device=" places N processes per device, which differs from
"--map-by device=" only in how many processes a device takes - no reason
at all for two enumerations that could disagree about what the devices
ARE, which is the same argument that put the enumeration in PMIx rather
than here.

So it moves to rmaps/base as a small API: begin a node, ask how many
processes it can take, ask what the j-th is placed against, record on a
proc which devices it was given, end.  A caller placing more than one
process per device simply asks for the same index more than once.

"begin" also owns the two checks that have to happen before any process
is placed on a node - that the binding is not coarser than the devices
are local to, and whether every device shares one locality - because a
second caller getting either of those wrong would be a bug nobody would
look for.

No behavior change is intended and none was observed: the offline
harness produces byte-identical maps across all 1777 cases.
